requirements.txt
# requirements.txt
python-telegram-bot==20.7
fastapi==0.109.2
uvicorn==0.27.1
